ecommerce-support-agent/pyproject.toml
[project]
name = "ecommerce-support-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Muhammad Aneeque", email = "aneeq.khatri@celeritasdigital.com" }
]
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
]

[project.scripts]
ecommerce-support-agent = "ecommerce_support_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
